requirements.txt
# 🚦 Smart Traffic AI System - Dependencies

# Core AI & Computer Vision
opencv-python==4.8.1.78          # Computer vision library
opencv-contrib-python==4.8.1.78  # Additional OpenCV modules
ultralytics==8.0.196             # YOLOv8 for object detection
torch==2.0.1                     # PyTorch for deep learning
torchvision==0.15.2               # Computer vision models
tensorflow==2.13.0               # TensorFlow alternative
numpy==1.24.3                    # Numerical computing
scipy==1.11.1                    # Scientific computing
scikit-learn==1.3.0              # Machine learning library
pandas==2.0.3                    # Data manipulation
matplotlib==3.7.2                # Plotting library
seaborn==0.12.2                  # Statistical visualization
pillow==10.0.0                   # Image processing

# Web Interface & API
fastapi==0.101.1                 # Modern web framework
uvicorn==0.23.2                  # ASGI server
jinja2==3.1.2                    # Template engine
websockets==11.0.3               # Real-time communication
requests==2.31.0                 # HTTP library
aiofiles==23.1.0                 # Async file operations

# Database & Storage
sqlalchemy==2.0.19               # Database ORM
alembic==1.11.1                  # Database migrations
redis==4.6.0                     # Caching and message broker
sqlite3                          # Built-in database (Python standard)

# IoT & Hardware Integration
pyserial==3.5                    # Serial communication
paho-mqtt==1.6.1                 # MQTT protocol
RPi.GPIO; platform_machine=="armv7l"  # Raspberry Pi GPIO (if on RPi)
adafruit-circuitpython-motor==3.4.8   # Motor control
gpiozero; platform_machine=="armv7l"   # GPIO control (Raspberry Pi)

# Time Series & Analytics
prophet==1.1.4                   # Time series forecasting
statsmodels==0.14.0              # Statistical modeling
plotly==5.15.0                   # Interactive plotting
dash==2.12.1                     # Interactive web apps
streamlit==1.25.0                # Quick web apps

# Utilities & Tools
python-dotenv==1.0.0             # Environment variables
pyyaml==6.0.1                    # YAML parsing
click==8.1.6                     # Command line interface
rich==13.4.2                     # Rich text and beautiful formatting
tqdm==4.65.0                     # Progress bars
schedule==1.2.0                  # Job scheduling
watchdog==3.0.0                  # File system monitoring

# Image & Video Processing
imageio==2.31.1                  # Image I/O
imageio-ffmpeg==0.4.8            # Video processing
av==10.0.0                       # Video/audio processing

# Security & Authentication
cryptography==41.0.3             # Cryptographic recipes
pyjwt==2.8.0                     # JSON Web Tokens
passlib==1.7.4                   # Password hashing
python-multipart==0.0.6          # Form data parsing

# Development & Testing
pytest==7.4.0                    # Testing framework
pytest-asyncio==0.21.1           # Async testing
black==23.7.0                    # Code formatter
flake8==6.0.0                    # Linting
mypy==1.5.1                      # Type checking
coverage==7.2.7                  # Code coverage

# Logging & Monitoring
loguru==0.7.0                    # Advanced logging
prometheus-client==0.17.1        # Metrics collection
psutil==5.9.5                    # System and process utilities

# Optional GPU Acceleration
# Uncomment if you have CUDA GPU
# torch-audio==2.0.2+cu118
# torchaudio==2.0.2+cu118

# Hardware-specific (uncomment as needed)
# picamera2; platform_machine=="armv7l"    # Raspberry Pi camera
# board; platform_machine=="armv7l"        # CircuitPython board
# busio; platform_machine=="armv7l"        # I2C/SPI communication

# Production deployment
gunicorn==21.2.0                 # Production WSGI server
docker==6.1.3                    # Docker Python SDK
